+++ FirebaseDynamicLinks/CHANGELOG.md
@@ -1,5 +1,5 @@
 # v8.8.0
-- [fixed] Firebase dynamic links with custom domain will only work if the custom domai has a trailing '/'. (#7087)
+- [fixed] Firebase dynamic links with custom domain will only work if the custom domain has a trailing '/'. (#7087)
 
 # v8.7.0
 - [added] Refactoring and adding helper class. (#8432)